requirements.txt
aiosqlite==0.16.1
asgiref==3.4.1
atlastk==0.12.10
click==8.0.1
fastapi==0.68.0
greenlet==1.1.1
gunicorn==20.0.4
h11==0.12.0
iso8601==0.1.16
pydantic==1.8.2
pypika-tortoise==0.1.1
python-decouple==3.5
pytz==2021.1
SQLAlchemy==1.4.22
starlette==0.14.2
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
uvicorn==0.15.0